implemented: Implemented a Query for Product on Stock Context

+++ src/Stoqui.Kernel.Domain/Extensions/QueryableExtensions.cs
@@ -2,14 +2,14 @@
 
 public static class QueryableExtensions
 {
-    public static IQueryable<T> Pagination<T>(this IQueryable<T> queryable, int page, ushort pageSize)
+    public static IQueryable<T> Pagination<T>(this IQueryable<T> queryable, int page, int pageSize)
     {
         return queryable
             .Skip((page - 1) * pageSize)
             .Take(pageSize);
     }
 
-    public static int GetTotalPage<T>(this IQueryable<T> queryable, ushort pageSize)
+    public static int GetTotalPage<T>(this IQueryable<T> queryable, int pageSize)
     {
         double totalItems = queryable.Count();
         return (int)Math.Ceiling( totalItems / pageSize);
